Extract BlockchainSimulationTestingCommons
