+++ src/test/scala/sigmastate/utxo/blockchain/BlockchainSimulationTestingCommons.scala
@@ -1,34 +1,42 @@
-package sigmastate.utxo
-
-import java.io.{File, FileWriter}
+package sigmastate.utxo.blockchain
 
 import org.ergoplatform
 import org.ergoplatform._
-import org.scalacheck.Gen
-import scorex.crypto.authds.avltree.batch.{BatchAVLProver, Insert, Remove}
 import scorex.crypto.authds.{ADDigest, ADKey, ADValue}
+import scorex.crypto.authds.avltree.batch.{BatchAVLProver, Insert, Remove}
 import scorex.crypto.hash.{Blake2b256, Digest32}
-import scorex.util._
+import sigmastate.{AvlTreeData, GE}
 import sigmastate.Values.LongConstant
 import sigmastate.eval.IRContext
-import sigmastate.helpers.{ContextEnrichingTestProvingInterpreter, ErgoLikeTestProvingInterpreter, ErgoTransactionValidator, SigmaTestingCommons}
-import sigmastate.interpreter.ContextExtension
-import sigmastate.interpreter.Interpreter.{ScriptNameProp, emptyEnv}
-import sigmastate.{AvlTreeData, GE}
+import sigmastate.helpers.{ErgoLikeTestProvingInterpreter, ErgoTransactionValidator, SigmaTestingCommons}
 
-import scala.annotation.tailrec
-import scala.collection.concurrent.TrieMap
 import scala.collection.mutable
 import scala.util.Try
+import scorex.util._
+import sigmastate.interpreter.ContextExtension
+import sigmastate.interpreter.Interpreter.{ScriptNameProp, emptyEnv}
+import sigmastate.utxo.blockchain.BlockchainSimulationTestingCommons.{FullBlock, ValidationState, heightReg}
 
+import scala.annotation.tailrec
 
-class BlockchainSimulationSpecification extends SigmaTestingCommons {
+trait BlockchainSimulationTestingCommons extends SigmaTestingCommons {
 
-  import BlockchainSimulationSpecification._
+  private val windowSize = 10
 
-  implicit lazy val IR = new TestingIRContext
+  @tailrec
+  protected final def checkState(state: ValidationState,
+                 miner: ErgoLikeTestProvingInterpreter,
+                 currentLevel: Int,
+                 limit: Int): Unit = currentLevel match {
+    case i if i >= limit => ()
+    case _ =>
+      val block = generateBlock(state, miner, currentLevel)
+      val updStateTry = state.applyBlock(block)
+      updStateTry.isSuccess shouldBe true
+      checkState(updStateTry.get, miner, currentLevel + 1, limit)
+  }
 
-  def generateBlock(state: ValidationState, miner: ErgoLikeTestProvingInterpreter, height: Int): Block = {
+  protected def generateBlock(state: ValidationState, miner: ErgoLikeTestProvingInterpreter, height: Int): FullBlock = {
     val minerPubKey = miner.dlogSecrets.head.publicImage
     val boxesToSpend = state.boxesReader.byHeightRegValue(height)
 
@@ -51,101 +59,20 @@ class BlockchainSimulationSpecification extends SigmaTestingCommons {
       tx.toSigned(IndexedSeq(proverResult))
     }.toIndexedSeq.ensuring(_.nonEmpty, s"Failed to create txs from boxes $boxesToSpend at height $height")
 
-    Block(txs, minerPubKey.pkBytes)
+    FullBlock(txs, minerPubKey.pkBytes)
   }
 
-  property("apply one valid block") {
-    val state = ValidationState.initialState()
-    val miner = new ErgoLikeTestProvingInterpreter()
-    val block = generateBlock(state, miner, 0)
-    val updStateTry = state.applyBlock(block)
-    updStateTry.isSuccess shouldBe true
-  }
-
-  property("too costly block") {
-    val state = ValidationState.initialState()
-    val miner = new ErgoLikeTestProvingInterpreter()
-    val block = generateBlock(state, miner, 0)
-    val updStateTry = state.applyBlock(block, maxCost = 1)
-    updStateTry.isSuccess shouldBe false
-  }
-
-  property("apply many blocks") {
-    val state = ValidationState.initialState()
-    val miner = new ErgoLikeTestProvingInterpreter()
-
-    @tailrec
-    def checkState(state: ValidationState,
-                   miner: ErgoLikeTestProvingInterpreter,
-                   currentLevel: Int,
-                   limit: Int): Unit = currentLevel match {
-      case i if i >= limit => ()
-      case _ =>
-        val block = generateBlock(state, miner, currentLevel)
-        val updStateTry = state.applyBlock(block)
-        updStateTry.isSuccess shouldBe true
-        checkState(updStateTry.get, miner, currentLevel + 1, limit)
-    }
-
-    val randomDeepness = Gen.chooseNum(10, 20).sample.getOrElse(15)
-    checkState(state, miner, 0, randomDeepness)
-  }
-
-  ignore(s"benchmarking applying many blocks (!!! ignored)") {
-    val results = new TrieMap[Int, Long]
-
-    def bench(numberOfBlocks: Int): Unit = {
-
-      val state = ValidationState.initialState()
-      val miner = new ContextEnrichingTestProvingInterpreter()
-
-      val (_, time) = (0 until numberOfBlocks).foldLeft(state -> 0L) { case ((s, timeAcc), h) =>
-        val b = generateBlock(state, miner, h)
-
-        val t0 = System.currentTimeMillis()
-        val updStateTry = s.applyBlock(b)
-        val t = System.currentTimeMillis()
-
-        updStateTry shouldBe 'success
-        updStateTry.get -> (timeAcc + (t - t0))
-      }
-
-      println(s"Total time for $numberOfBlocks blocks: $time ms")
-      results.put(numberOfBlocks, time)
-    }
-
-    bench(100)
-    bench(200)
-    bench(300)
-    bench(400)
-
-    printResults(results.toMap)
-
-    def printResults(results: Map[Int, Long]): Unit = {
-      val file = new File("target/bench")
-      file.mkdirs()
-      val writer = new FileWriter(s"target/bench/result.csv", false)
-      val sorted = results.toList.sortBy { case (i, _) => i }
-      val header = sorted.map(_._1).mkString(",")
-      writer.write(s"$header\n")
-      val values = sorted.map(_._2).mkString(",")
-      writer.write(s"$values\n")
-      writer.flush()
-      writer.close()
-    }
-  }
 }
 
-object BlockchainSimulationSpecification {
-  private lazy val hash = Blake2b256
+object BlockchainSimulationTestingCommons {
 
   val heightReg = ErgoBox.nonMandatoryRegisters.head
 
-  val windowSize = 10
+
 
   val MaxBlockCost = 700000
 
-  case class Block(txs: IndexedSeq[ErgoLikeTransaction], minerPubkey: Array[Byte])
+  case class FullBlock(txs: IndexedSeq[ErgoLikeTransaction], minerPubkey: Array[Byte])
 
   class InMemoryErgoBoxReader(prover: ValidationState.BatchProver) extends ErgoBoxReader {
     private type KeyType = mutable.WrappedArray.ofByte
@@ -170,7 +97,7 @@ object BlockchainSimulationSpecification {
 
     def allIds: Iterable[KeyType] = boxes.keys
 
-    def applyBlock(block: Block): Unit = {
+    def applyBlock(block: FullBlock): Unit = {
       val toRemove = block.txs.flatMap(_.inputs).map(_.boxId)
       toRemove.foreach(k => prover.performOneOperation(Remove(k)))
       toRemove.foreach(k => boxes.remove(getKey(k)))
@@ -189,7 +116,7 @@ object BlockchainSimulationSpecification {
   case class ValidationState(state: BlockchainState, boxesReader: InMemoryErgoBoxReader)(implicit IR: IRContext) {
     val validator = new ErgoTransactionValidator
 
-    def applyBlock(block: Block, maxCost: Int = MaxBlockCost): Try[ValidationState] = Try {
+    def applyBlock(block: FullBlock, maxCost: Int = MaxBlockCost): Try[ValidationState] = Try {
       val height = state.currentHeight + 1
 
       val blockCost = block.txs.foldLeft(0L) { case (accCost, tx) =>
@@ -212,16 +139,16 @@ object BlockchainSimulationSpecification {
   object ValidationState {
     type BatchProver = BatchAVLProver[Digest32, Blake2b256.type]
 
-    val initBlock = Block(
-      (0 until windowSize).map { i =>
-        val txId = hash.hash(i.toString.getBytes ++ scala.util.Random.nextString(12).getBytes).toModifierId
+    val initBlock = FullBlock(
+      (0 until 10).map { i =>
+        val txId = Blake2b256.hash(i.toString.getBytes ++ scala.util.Random.nextString(12).getBytes).toModifierId
         val boxes = (1 to 50).map(_ => ErgoBox(10, GE(Height, LongConstant(i)).toSigmaProp, 0, Seq(), Map(heightReg -> LongConstant(i)), txId))
         ergoplatform.ErgoLikeTransaction(IndexedSeq(), boxes)
       },
       ErgoLikeContext.dummyPubkey
     )
 
-    def initialState(block: Block = initBlock)(implicit IR: IRContext): ValidationState = {
+    def initialState(block: FullBlock = initBlock)(implicit IR: IRContext): ValidationState = {
       val keySize = 32
       val prover = new BatchProver(keySize, None)
 
@@ -235,5 +162,4 @@ object BlockchainSimulationSpecification {
       ValidationState(bs, boxReader).applyBlock(block).get
     }
   }
-
 }
